+++ src/eso_build_o_rama/__init__.py
@@ -1,0 +1,6 @@
+"""
+ESO Build-O-Rama
+A weekly scan of ESOLogs to publish new top-performing trial builds.
+"""
+
+__version__ = "0.1.0"